attempt to generate matrics with better build cache, by jules
Prompt:
build cache doesn't seem to work for the docker build github actions. Can you modernize to the latest best practices for github actions docker image builds (make sure to preserve support for interdependencies between the different builds in this repo) and set up everything for build caching?

Prompt 2:
note that ./test.sh has to be able to generate the new workflows syntax

For some reason it refuses to push a branch so I have downloaded the file.